+++ src/utils/invariant.ts
@@ -1,0 +1,6 @@
+export default function invariant(
+  condition: unknown,
+  error: string
+): asserts condition {
+  if (!condition) throw new Error(error);
+}