+++ type-definitions/ts-tests/map.ts
@@ -85,7 +85,7 @@ test('#get', () => {
 });
 
 test('#getIn', () => {
-  const result = Map({ a: 4, b: true }).getIn(['a' as const]);
+  const result = Map({ a: 4, b: true }).getIn(['a']);
 
   expect(result).type.toBeNumber();
 
@@ -100,9 +100,9 @@ test('#getIn', () => {
     ])
   ).type.toBeNumber();
 
-  // currently `RetrievePathReducer` does not work with anything else than `MapOf`
-  // TODO : fix this with a better type, it should be resolved to `number` (and not be marked as `fail`)
-  expect.fail(Map({ a: List([1]) }).getIn(['a' as const, 0])).type.toBeNumber();
+  expect(Map({ a: [1] }).getIn(['a' as const, 0])).type.toBeNumber();
+
+  expect(Map({ a: List([1]) }).getIn(['a' as const, 0])).type.toBeNumber();
 });
 
 test('#set', () => {